Add missing `purescript-strings` dependency
The build was working because some other dependency in `devDependencies`
would bring in `purescript-strings` transitively.
That's fine for this package, and it mostly works in the wild.

However, if you're depending on this package without also transitively
depending on `purescript-strings`, building this package fails.
It's rare, but it happens.

The failure can be reproduced before this commit by running:
```sh
$ bower install --production
$ pulp build
```
That will only install the `dependencies`.
And it will fail with not being able to find `Data.String`.
This is the behavior of what someone depending on this package would get.

After this commit, the above lines run without an error.